requirements.txt
aiohttp>=3.8.5
aiosignal>=1.3.1
aiounittest>=1.3.0
async-timeout>=4.0.3
attrs>=23.1.0
Babel>=2.9.1
botbuilder-core>=4.14.4
botbuilder-dialogs>=4.14.4
botbuilder-schema>=4.14.4
botframework-connector>=4.14.4
botframework-streaming>=4.14.4
certifi>=2023.7.22
cffi>=1.15.1
charset-normalizer>=3.2.0
cryptography>=41.0.4
datedelta>=1.4
emoji>=1.7.0
frozenlist>=1.4.0
grapheme>=0.6.0
idna>=3.4
isodate>=0.6.1
jsonpickle>=1.4.2
msal>=1.24.0
msrest>=0.6.21
multidict>=6.0.4
multipledispatch>=1.0.0
oauthlib>=3.2.2
pip>=23.2.1
prettyconf>=2.2.1
pycparser>=2.21
pytz>=2023.3.post1
recognizers-text>=1.0.2a2
recognizers-text-choice>=1.0.2a2
recognizers-text-date-time>=1.0.2a2
recognizers-text-number>=1.0.2a2
recognizers-text-number-with-unit>=1.0.2a2
regex>=2019.8.19
requests>=2.31.0
requests-oauthlib>=1.3.1
setuptools>=68.0.0
six>=1.16.0
topdesk>=0.0.17
ujson>=5.8.0
urllib3>=1.26.16
yarl>=1.9.2
sqlalchemy==2.0.21
pymysql==1.1.0
werkzeug==2.3.7